Move logic to separate files.
